Move importing logic to settings screen
