+++ app/src/main/java/journal/gratitude/com/gratitudejournal/ui/timeline/TimelineFragment.kt
@@ -196,48 +196,6 @@ class TimelineFragment : DaggerFragment() {
         }
     }
 
-
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        when (requestCode) {
-            IMPORT_CSV -> {
-                if (resultCode == RESULT_OK) {
-                    val uri = data?.data
-                    if (uri != null) {
-                        if (uri.scheme == "content") {
-                            val inputStream = activity?.contentResolver?.openInputStream(uri)
-                            if (inputStream != null) {
-                                importFromCsv(inputStream)
-                            } else {
-                                Crashlytics.logException(NullPointerException("inputStream is null, uri: $uri"))
-                                Toast.makeText(context, R.string.error_parsing, Toast.LENGTH_SHORT)
-                                    .show()
-                            }
-                        }
-                    } else {
-                        Crashlytics.log("URI was null when receiving file")
-                        Toast.makeText(context, R.string.file_not_csv, Toast.LENGTH_SHORT).show()
-                    }
-                }
-            }
-        }
-    }
-
-    private fun importFromCsv(inputStream: InputStream) {
-        // parse file to get List<Entry>
-        try {
-            val csvReader = CSVReaderImpl(inputStream.bufferedReader())
-            val entries = parseCsv(csvReader)
-            viewModel.addEntries(entries)
-            firebaseAnalytics.logEvent(IMPORTED_DATA_SUCCESS, null)
-        } catch (exception: Exception) {
-            firebaseAnalytics.logEvent(IMPORTING_BACKUP_ERROR, null)
-            Crashlytics.logException(exception)
-
-            Toast.makeText(context, R.string.error_parsing, Toast.LENGTH_SHORT).show()
-        }
-    }
-
     private fun openContactForm() {
         firebaseAnalytics.logEvent(OPENED_CONTACT_FORM, null)
